requirements.txt
psycopg2-binary==2.9.11
pymongo==4.8.0
python-dotenv==1.0.1